backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
gunicorn==21.2.0
python-dotenv==1.0.1
Pillow==10.3.0
pytest==8.0.0
black==24.1.1
flake8==7.0.0
